[Abhishek] - Added text field data for some details

+++ app/src/main/java/com/aby/c0769778_w2020_mad3125_midterm/model/CRACustomer.java
@@ -125,7 +125,7 @@ public class CRACustomer implements Parcelable
         this.birthDate = birthDate;
     }
 
-    public Float getGrossIncome() {
+    public double getGrossIncome() {
         return grossIncome;
     }
 
@@ -133,7 +133,7 @@ public class CRACustomer implements Parcelable
         this.grossIncome = grossIncome;
     }
 
-    public Float getRrspContributed() {
+    public double getRrspContributed() {
         return rrspContributed;
     }
